[TM-812] Roll back some earlier changes that aren't necessary now that the UUID is being generated on the client.
