+++ src/pages/entity/[entityName]/edit/[uuid]/index.page.tsx
@@ -1,7 +1,7 @@
 import { useT } from "@transifex/react";
 import { notFound } from "next/navigation";
 import { useRouter } from "next/router";
-import { useMemo, useRef } from "react";
+import { useMemo } from "react";
 
 import WizardForm from "@/components/extensive/WizardForm";
 import BackgroundLayout from "@/components/generic/Layout/BackgroundLayout";
@@ -40,13 +40,7 @@ const EditEntityPage = () => {
   });
   const entity = entityData?.data || {}; //Do not abuse this since forms should stay entity agnostic!
 
-  const {
-    mutate: updateEntity,
-    data: updateData,
-    error,
-    isSuccess,
-    isLoading: isUpdating
-  } = usePutV2FormsENTITYUUID({});
+  const { mutate: updateEntity, error, isSuccess, isLoading: isUpdating } = usePutV2FormsENTITYUUID({});
   const { mutate: submitEntity, isLoading: isSubmitting } = usePutV2FormsENTITYUUIDSubmit({
     onSuccess() {
       if (mode === "edit" || mode?.includes("provide-feedback")) {
@@ -68,18 +62,7 @@ const EditEntityPage = () => {
   //@ts-ignore
   const formData = (data?.data || {}) as GetV2FormsENTITYUUIDResponse;
 
-  // updateData will briefly be null while the update is in progress, and during that period we want
-  // to make sure to keep using the previous value, so we need to use a ref to track that state
-  const contentDataRef = useRef<GetV2FormsENTITYUUIDResponse | null>(null);
-  if (contentDataRef.current == null && formData?.answers != null) {
-    contentDataRef.current = formData;
-  } else if (updateData != null) {
-    // @ts-ignore
-    contentDataRef.current = updateData?.data;
-  }
-
-  const feedbackFields =
-    contentDataRef.current?.update_request?.feedback_fields ?? contentDataRef.current?.feedback_fields ?? [];
+  const feedbackFields = formData?.update_request?.feedback_fields ?? formData?.feedback_fields ?? [];
 
   const formSteps = useGetCustomFormSteps(
     formData.form,
@@ -91,8 +74,8 @@ const EditEntityPage = () => {
     mode?.includes("provide-feedback") ? feedbackFields : undefined
   );
 
-  const values = useNormalizedFormDefaultValue(
-    contentDataRef.current?.update_request?.content ?? contentDataRef.current?.answers,
+  const defaultValues = useNormalizedFormDefaultValue(
+    formData?.update_request?.content ?? formData?.answers,
     formSteps,
     entity.migrated
   );
@@ -146,7 +129,7 @@ const EditEntityPage = () => {
             })
           }
           submitButtonDisable={isSubmitting}
-          values={values}
+          defaultValues={defaultValues}
           title={formTitle}
           tabOptions={{
             markDone: true,
